feat: 데이터 수집 및 임의 전처리와 일괄 저장 처리
