+++ src/main/java/org/jun/saemangeum/process/application/service/ContentCollectService.java
@@ -1,0 +1,27 @@
+package org.jun.saemangeum.process.application.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.jun.saemangeum.global.persistence.repository.ContentRepository;
+import org.jun.saemangeum.process.application.collect.base.Refiner;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class ContentCollectService {
+
+    private final List<Refiner> refiners;
+    private final ContentRepository contentRepository;
+    // ...+ 전처리용 서비스 로직 추가?
+
+    @Transactional
+    public void collectAndSave() {
+        refiners.stream()
+                .peek(e -> log.info("각 구현체들 전처리 예정"))
+                .forEach(refiner -> contentRepository.saveAll(refiner.refine()));
+    }
+}
